refactor:  Hit()의 매개변수 타입 변경(int -> double)
적 체력 감소를 제대로 보여주기 위해 double 형 필요

+++ TextRpg/Monster.cs
@@ -102,7 +102,7 @@ namespace TextRpg
             return $"Lv.{Level} {Name} HP: {CurrentHP}/{MaxHP}";
         }
 
-        public void Hit(int damage)
+        public void Hit(double damage)
         {
             _currentHP = Math.Max(0, _currentHP - damage);
         }
